components/esp-opus/dnn/torch/osce/requirements.txt
pyyaml==6.0.1
torch==2.0.1
numpy==1.25.2
scipy==1.11.2
pesq==0.0.4
gitpython==3.1.41
matplotlib==3.7.3
torchaudio==2.0.2
tqdm==4.66.1
